+++ test/e2e/suite/command/command_suite_test.go
@@ -13,7 +13,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package command_test
+package command
 
 import (
 	"testing"
