+++ test/e2e/suite/command/version.go
@@ -13,7 +13,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package command_test
+package command
 
 import (
 	"os/exec"
